Moved mdns_rtpmidi into the library
