+++ src/CMakeLists.txt
@@ -1,10 +1,10 @@
 add_executable(
   rtpmidid-daemon  
-  mdns_rtpmidi.cpp aseq.cpp stringpp.cpp
+  aseq.cpp stringpp.cpp
   main.cpp config.cpp rtpmidid.cpp
   control_socket.cpp
 )
 
-target_link_libraries(rtpmidid-daemon rtpmidid-static -lasound -lfmt -lavahi-common -lavahi-client)
+target_link_libraries(rtpmidid-daemon rtpmidid-static -lasound -lfmt)
 
 set_target_properties(rtpmidid-daemon PROPERTIES OUTPUT_NAME rtpmidid)
